% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_hh}
\alias{get_hh}
\title{Get household data for a single wave}
\usage{
get_hh(dir, wave, vars = c(), long = TRUE)
}
\arguments{
\item{dir}{The path to the directory containing the subfolders for each
wave. This is named "stata13_se" when downloading from UKDS.}

\item{wave}{A whole number indicating the desired wave of Understanding
Society data.}

\item{vars}{A character vector containing the desired variables.
Variables should be specified without the "w_" prefix.
If an empty vector is provided, all variables will be returned.}

\item{long}{The format of the output.
If TRUE, all variable "w_" prefixes are removed and a column is added
indicating the numerical wave of the data.}
}
\value{
A tibble containing household response data for the desired wave
and variables.
}
\description{
Gets household response data for a particular wave and set of variables.
}
\examples{
\dontrun{
get_hh(10, vars=c("agechy_dv", "hidp"))
get_hh(9)
}

}
