% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_ind}
\alias{get_ind}
\title{Get individual data for a single wave}
\usage{
get_ind(dir, wave, vars = c(), long = TRUE)
}
\arguments{
\item{dir}{The path to the directory containing the subfolders for each
wave. This is named "stata13_se" when downloading from UKDS.}

\item{wave}{A whole number indicating the desired wave of Understanding
Society data.}

\item{vars}{A character vector containing the desired variables.
Variables should be specified without the "w_" prefix and "pidp" is
included by default.
If an empty vector is provided, all variables will be returned.}

\item{long}{The format of the output.
If TRUE, all variable "w_" prefixes are removed and a column is added
indicating the numerical wave of the data.}
}
\value{
A tibble containing individual response data for the desired wave
and variables.
}
\description{
Gets individual response data for a particular wave and set of variables.
}
\examples{
\dontrun{
get_ind(10, vars="envhabit1")
get_ind(9)
}

}
