% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_ind_multiwave}
\alias{get_ind_multiwave}
\title{Get individual data for multiple waves}
\usage{
get_ind_multiwave(dir, waves, vars, long = TRUE)
}
\arguments{
\item{dir}{The path to the directory containing the subfolders for each
wave. This is named "stata13_se" when downloading from UKDS.}

\item{waves}{A vector of whole numbers indicating the desired waves of
Understanding Society data.}

\item{vars}{A character vector containing the desired variables.
Variables should appear in all specified waves.
Variables should be specified without the "w_" prefix and "pidp" is
included by default.
If an empty vector is provided, all variables will be returned.}

\item{long}{The format of the output.
If TRUE, data is returned in "long" format with all variable "w_"
prefixes removed and with a column to indicate the numerical wave of the
data.
Otherwise, data is returned in "wide" format with a column for each wave
and variable.}
}
\value{
A tibble containing individual response data for the desired
waves and variables.
}
\description{
Gets individual response data for multiple waves.
}
\examples{
\dontrun{
get_ind_multiwave(c(4, 10), vars="envhabit1")
get_ind(c(4, 10), vars=c("envhabit1", "envhabit2"), long=FALSE)
}

}
