% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{make_na}
\alias{make_na}
\title{Replace special values with NA}
\usage{
make_na(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
The same dataframe with Understanding Society's missing values
replaced with NA.
}
\description{
Replaces Understanding Society's special "missing" values with NA.
}
\examples{
test_data = data.frame(a = c(532, -2), b = c(-9, 40))
clean_data = make_na(test_data)
clean_data

}
